Replace `util_libc` and `sanitizer` modules with `impl_utils` macro
